Obtain applied interval depth from genesis txs byte.

+++ include/bitcoin/database/settings.hpp
@@ -43,7 +43,7 @@ struct BCD_API settings
     bool turbo{ false };
 
     /// Depth of electrum merkle tree interval caching.
-    uint8_t interval_depth{ max_uint8 };
+    uint16_t interval_depth{ max_uint8 };
 
     /// Path to the database directory.
     std::filesystem::path path{ "bitcoin" };
